requirements.txt
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
playwright==1.43.0
openai==1.40.0
httpx==0.27.2
